src/requirements.txt
iambic-core==0.11.97
cloudaux-lite==1.0.0
click==8.1.7
PyYAML==6.0.1
marshmallow==3.21.1
retry==0.9.2
pyjwt==2.8.0
requests==2.31.0
# cryptography==42.0.5  Handled by iambic-core
botocore==1.34.71
boto3==1.34.71
slack-sdk==3.27.1
deepdiff==6.7.1
